+++ pyearth/_pruning.pyx
@@ -5,6 +5,7 @@
 # cython: profile = False
 
 from ._record cimport PruningPassIteration
+from ._types import FLOAT
 from ._util cimport gcv, apply_weights_2d
 import numpy as np
 from scipy.linalg import lstsq
@@ -29,7 +30,7 @@ cdef class PruningPasser:
         self.sample_weight = sample_weight
         self.verbose = verbose
         self.basis = basis
-        self.B = np.empty(shape=(self.m, len(self.basis) + 1), dtype=np.float)
+        self.B = np.empty(shape=(self.m, len(self.basis) + 1), dtype=FLOAT)
         self.penalty = kwargs.get('penalty', 3.0)
         if sample_weight.shape[1] == 1:
             y_avg = np.average(self.y, weights=sample_weight[:,0], axis=0)
